feat(app): add movie data adapter
