+++ src/components/MoviesCard/MoviesCard.jsx
@@ -2,18 +2,22 @@ import { useState } from 'react';
 import { useLocation } from 'react-router-dom';
 
 import { timeConvertor } from './../../utils/timeConvertor.js';
-
-import './MoviesCard.css';
+import mainApi from './../../utils/MainApi.js';
 import { movieApiURL } from './../../constants/constants.js';
 
+import './MoviesCard.css';
+
 export default function MoviesCard({ nameRU, movieId, duration, image, isSaved, className: classList = '' }) {
-  const [cardSaved, setCardSaved] = useState(isSaved);
+  const [isMovieSaved, setIsMovieSaved] = useState(isSaved);
   const imageURL = `${movieApiURL}${image.url}`;
 
   const location = useLocation();
 
   const saveMovieHandler = () => {
-    setCardSaved(!cardSaved);
+    setIsMovieSaved(!isMovieSaved);
+    if (!isMovieSaved) {
+      mainApi.createMovie();
+    }
   };
 
   const deleteMovieHandler = () => {
@@ -29,7 +33,7 @@ export default function MoviesCard({ nameRU, movieId, duration, image, isSaved, 
       </div>
       {location.pathname === '/movies' && (
         <button
-          className={`movie-card__btn movie-card__btn_type${cardSaved ? '_saved' : '_unsaved'}`}
+          className={`movie-card__btn movie-card__btn_type${isMovieSaved ? '_saved' : '_unsaved'}`}
           onClick={saveMovieHandler}
         >
           Сохранить
